requirements.txt
# Requirements

anthropic>=0.25.0
openai>=1.0.0
chromadb>=0.4.0
pydantic>=2.0.0
httpx>=0.25.0
beautifulsoup4>=4.12.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
python-dotenv>=1.0.0
rich>=13.0.0

# Backend API (optional)
fastapi>=0.100.0
uvicorn>=0.23.0
